add basic route handlers
