--- src/app/api/hello/route.ts
@@ -1,10 +1,0 @@
-import { NextRequest, NextResponse } from 'next/server'
-
-export async function GET(request: NextRequest) {
-  const greeting = 'Hello World!!'
-  const json = {
-    greeting,
-  }
-
-  return NextResponse.json(json)
-}
